chore: refactor python script
